requirements.txt
pandas
psycopg2-binary
requests
google-ads
facebook-business
mailchimp-marketing
python-dotenv
